# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EMKA Enterprise Control Core",
    "description": "Autonomous Intelligence Engine & System Triage Control Hub",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/triage": {
      "post": {
        "tags": [
          "Incident Triage Engine"
        ],
        "summary": "Process Incident Stream",
        "operationId": "process_incident_stream_api_v1_triage_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/history": {
      "get": {
        "tags": [
          "Incident Triage Engine"
        ],
        "summary": "Get Triage History",
        "operationId": "get_triage_history_api_v1_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/incidents/escalate": {
      "post": {
        "tags": [
          "Incident Triage Engine"
        ],
        "summary": "Escalate Incident",
        "operationId": "escalate_incident_api_v1_incidents_escalate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentEscalateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/logs/{log_id}/escalate": {
      "post": {
        "tags": [
          "Incident Triage Engine"
        ],
        "summary": "Escalate Compliance Node",
        "operationId": "escalate_compliance_node_api_v1_logs__log_id__escalate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/logs/{log_id}": {
      "delete": {
        "tags": [
          "Incident Triage Engine"
        ],
        "summary": "Purge Compliance Node",
        "operationId": "purge_compliance_node_api_v1_logs__log_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnvironmentContext": {
        "properties": {
          "emails": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Emails"
          },
          "local_files": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Local Files"
          }
        },
        "type": "object",
        "required": [
          "emails",
          "local_files"
        ],
        "title": "EnvironmentContext"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentEscalateRequest": {
        "properties": {
          "incident_id": {
            "type": "integer",
            "title": "Incident Id"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity",
            "default": "HIGH"
          }
        },
        "type": "object",
        "required": [
          "incident_id"
        ],
        "title": "IncidentEscalateRequest"
      },
...
      "TriageRequest": {
        "properties": {
          "task_name": {
            "type": "string",
            "title": "Task Name"
          },
          "context": {
            "$ref": "#/components/schemas/EnvironmentContext"
          }
        },
        "type": "object",
        "required": [
          "task_name",
          "context"
        ],
        "title": "TriageRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}